requirements.txt
httpx[http2]>=0.27
beautifulsoup4>=4.12
lxml>=5.1
selenium>=4.18
playwright>=1.42
pyyaml>=6.0
python-dotenv>=1.0
ollama>=0.4
jinja2>=3.1
rich>=13.7
